pyproject.toml
[project]
name = "itsg"
version = "0.0.1"
description = ""
authors = [
    { name = "Caroline Mimeault", email = "carolinemimeault@gmail.com" },
    { name = "Mike Williamson", email = "mike@korora.ca" },
]
dependencies = [
    "fastapi>=0.92.0",
    "uvicorn[standard]>=0.20.0",
    "strawberry-graphql[fastapi]>=0.158.0",
]
requires-python = ">=3.9"
license = { text = "MIT" }
[project.optional-dependencies]

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = ["httpx>=0.23.3", "pytest-cov>=4.0.0", "ruff>=0.0.247"]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

# to restrict tests to the ones in the app
[tool.pytest.ini_options]
testpaths = "app"

[tool.pdm.scripts]
start = "uvicorn app.main:app --reload"
test = "pytest --cov=app"
